+++ applovin_max/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Versions
 
-## x.x.x
+## 2.4.1
 * Depends on Android SDK v11.7.1 and iOS SDK v11.7.1.
 * Fix `MaxAdView` widget creation error on Android. Fixes https://github.com/AppLovin/AppLovin-MAX-Flutter/issues/76.
 ## 2.4.0